bus-realtime-processor/requirements.txt
pyspark==3.4.0
kafka-python
networkx==3.1
pandas==2.0.3
numpy==1.24.4